fix: fix stat on linux
