+++ internal/shared/fs/local.go
@@ -2,8 +2,6 @@ package fs
 
 import (
 	"os"
-	"syscall"
-	"time"
 )
 
 type Local struct {
@@ -25,20 +23,4 @@ func (*Local) Mkdir(path string) error {
 
 func (*Local) CreateFile(path string, content []byte) error {
 	return os.WriteFile(path, content, 0755)
-}
-
-func (*Local) Stat(path string) (*FileInfo, error) {
-	stat, err := os.Stat(path)
-	if err != nil {
-		return nil, err
-	}
-
-	sysStat := stat.Sys().(*syscall.Stat_t)
-	ctime := time.Unix(sysStat.Birthtimespec.Sec, sysStat.Birthtimespec.Nsec)
-
-	return &FileInfo{
-		Path:      path,
-		CreatedAt: ctime,
-		UpdatedAt: stat.ModTime(),
-	}, nil
 }
